+++ src/nanotron/trainer.py
@@ -564,9 +564,6 @@ class DistributedTrainer:
             self.config, self.parallel_context, self.unwrapped_model, self.grad_accumulator, self.optimizer
         )
 
-        if dist.get_rank() == 0:
-            assert 1 == 1
-
         # Apply gradient
         self.optimizer.step()
         self.optimizer.zero_grad()
@@ -594,8 +591,8 @@ class DistributedTrainer:
         # if dist.get_rank() == 0 and constants._NOT_BWD_ASYNC_OPS:
         #     assert 1 == 1
 
-        assert len(not_finished) == 0, f"AsyncCommBucket._copy_async_op: {not_finished}"
-        assert len(AsyncCommBucket._async_op) == 0, f"AsyncCommBucket._async_op: {AsyncCommBucket._async_op}"
+        # assert len(not_finished) == 0, f"AsyncCommBucket._copy_async_op: {not_finished}"
+        # assert len(AsyncCommBucket._async_op) == 0, f"AsyncCommBucket._async_op: {AsyncCommBucket._async_op}"
         AsyncCommBucket.clear_all()
 
         return outputs, loss_avg
